OBSインポート機能の改善: ボタンラベルを変更し、OBS Studioに関する文言を修正
