+++ app/components/windows/SceneTransitions.vue
@@ -5,7 +5,6 @@
     :content-styles="{ padding: 0 }"
     :show-cancel="false"
     :done-handler="done"
-    :customControls="true"
   >
     <div slot="content">
       <div v-if="!transitionsEnabled" class="transition-blank">
@@ -111,11 +110,6 @@
           </div>
         </div>
       </modal>
-    </div>
-    <div slot="controls">
-      <button class="button button--secondary" @click="importFromOBS" :disabled="!canImportFromOBS">
-        {{ $t('onboarding.importFromObs') }}
-      </button>
     </div>
   </modal-layout>
 </template>
